Render question answers from sessionStorage
Renders the stored answers in to the respective answer "slot", marked by
the `data-question` attribute.

We may want to add placeholders at some point, but I just left the
existing defaults for now as I'm not sure how modern you need the
browser to be to support what we're doing.